+++ lib/src/components/_internal_components.dart
@@ -48,7 +48,7 @@ class LiveTimeIndicator extends StatefulWidget {
 
   /// Widget to display tile line according to current time.
   const LiveTimeIndicator(
-      {Key? key,
+      {super.key,
       required this.width,
       required this.height,
       required this.timeLineWidth,
@@ -57,8 +57,7 @@ class LiveTimeIndicator extends StatefulWidget {
       required this.startHour,
       this.endHour = Constants.hoursADay,
       this.onlyShowToday = false,
-      this.date})
-      : super(key: key);
+      this.date});
 
   @override
   _LiveTimeIndicatorState createState() => _LiveTimeIndicatorState();
@@ -232,7 +231,7 @@ class TimeLine extends StatefulWidget {
 
   /// Time line to display time at left side of day or week view.
   const TimeLine({
-    Key? key,
+    super.key,
     required this.timeLineWidth,
     required this.hourHeight,
     required this.height,
@@ -246,7 +245,7 @@ class TimeLine extends StatefulWidget {
     this.showQuarterHours = false,
     this.endHour = Constants.hoursADay,
     this.showMidnightHour = false,
-  }) : super(key: key);
+  });
 
   @override
   State<TimeLine> createState() => _TimeLineState();
@@ -448,7 +447,7 @@ class EventGenerator<T extends Object?> extends StatelessWidget {
 
   /// A widget that display event tiles in day/week view.
   const EventGenerator({
-    Key? key,
+    super.key,
     required this.height,
     required this.width,
     required this.events,
@@ -462,7 +461,7 @@ class EventGenerator<T extends Object?> extends StatelessWidget {
     required this.scrollNotifier,
     required this.onTileDoubleTap,
     this.endHour = Constants.hoursADay,
-  }) : super(key: key);
+  });
 
   /// Arrange events and returns list of [Widget] that displays event
   /// tile on display area. This method uses [eventArranger] to get position
@@ -588,7 +587,7 @@ class PressDetector extends StatelessWidget {
 
   /// A widget that display event tiles in day/week view.
   const PressDetector({
-    Key? key,
+    super.key,
     required this.height,
     required this.width,
     required this.heightPerMinute,
@@ -597,7 +596,7 @@ class PressDetector extends StatelessWidget {
     required this.onDateTap,
     required this.minuteSlotSize,
     required this.startHour,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {