+++ lib/src/components/headers/day_page_header.dart
@@ -2,42 +2,26 @@
 // Use of this source code is governed by a MIT-style license
 // that can be found in the LICENSE file.
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-
 import '../../../calendar_view.dart';
-import '../../constants.dart';
 
 /// A header widget to display on day view.
 class DayPageHeader extends CalendarPageHeader {
   /// A header widget to display on day view.
   const DayPageHeader({
-    Key? key,
-    VoidCallback? onNextDay,
-    bool showNextIcon = true,
-    AsyncCallback? onTitleTapped,
-    VoidCallback? onPreviousDay,
-    bool showPreviousIcon = true,
+    super.key,
+    super.onNextDay,
+    super.showNextIcon,
+    super.onTitleTapped,
+    super.onPreviousDay,
+    super.showPreviousIcon,
     StringProvider? dateStringBuilder,
-    required DateTime date,
-    @Deprecated("Use HeaderStyle to provide icon color") Color? iconColor,
-    @Deprecated("Use HeaderStyle to provide background")
-    Color backgroundColor = Constants.headerBackground,
-    HeaderStyle headerStyle = const HeaderStyle(),
+    required super.date,
+    @Deprecated("Use HeaderStyle to provide icon color") super.iconColor,
+    @Deprecated("Use HeaderStyle to provide background") super.backgroundColor,
+    super.headerStyle,
   }) : super(
-          key: key,
-          date: date,
-          // ignore_for_file: deprecated_member_use_from_same_package
-          backgroundColor: backgroundColor,
-          iconColor: iconColor,
-          onNextDay: onNextDay,
-          showNextIcon: showNextIcon,
-          onPreviousDay: onPreviousDay,
-          showPreviousIcon: showPreviousIcon,
-          onTitleTapped: onTitleTapped,
           dateStringBuilder:
               dateStringBuilder ?? DayPageHeader._dayStringBuilder,
-          headerStyle: headerStyle,
         );
 
   static String _dayStringBuilder(DateTime date, {DateTime? secondaryDate}) =>